refactor: separate DialogContent to new file, InviteMemberDialogContent.tsx
