+++ src/components/page/MembersPage.tsx
@@ -1,6 +1,6 @@
 import { MenuMeatballIcon, PersonCircleIcon } from '@/assets/images/icons';
 import { Button, Card, Icon } from '@/components/ui';
-import { InviteMemberButton } from '@/features/groups/components/InviteMemberButton';
+import { InviteMemberDialog } from '@/features/groups/components/InviteMemberDialog';
 import { fetchUserList } from '@/lib/api/user/server';
 import { IUser } from '@/types/definition';
 
@@ -18,7 +18,7 @@ export const MembersPage = async ({ id }: IMembersPageProps) => {
           <Button className="flex justify-center items-center w-12 h-12">
             <Icon icon={MenuMeatballIcon} size={5} />
           </Button>
-          <InviteMemberButton />
+          <InviteMemberDialog />
         </div>
       </div>
       <div className="flex flex-col gap-2 pb-1">
